# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Thailand Enterprise Accounting System API
  description: |
    REST API สำหรับระบบบัญชี enterprise ของไทย

    **Authentication:** API Key + HMAC signature, OR OAuth2 client_credentials
    **Idempotency:** ทุก POST/PUT ต้องมี Idempotency-Key header (UUID v4)
    **Rate limit:** 1,000 req/min per API key (default)
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
servers:
  # Host root only — paths carry their own prefix. BFF/JWT routes are bare
  # (e.g. /quotations); the external API key surface is under /api/v1/*; MCP at /mcp.
  - url: https://api.accounting.example.com
    description: Production
  - url: https://api-staging.accounting.example.com
    description: Staging
security:
  - ApiKeyAuth: []
  - OAuth2: [read, write]
paths:
...
  /tax-filings/cit/profile:
    get:
      summary: CIT profile for a fiscal year — auto-SME test + ภ.ง.ด.50 inputs
      description: >
        SME = paid-up capital ≤ ฿5M AND revenue ≤ ฿30M (null capital → General, never silently SME).
        lossCarryIn = loss carry-forward per ม.65ตรี(12) over the per-FY store (5-period expiry,
        oldest-first consumption).
      tags: [Tax Filings]
      parameters:
        - name: year
          in: query
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: CIT profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  fiscalYear: { type: integer }
                  paidUpCapital: { type: number, nullable: true }
                  revenueFullYear: { type: number }
                  isSme: { type: boolean }
                  adjustmentsTotal: { type: number }
                  lossCarryIn: { type: number }
                  accountingNetProfit: { type: number }
...
  /reports/pnd30/file:
    get:
      summary: Generate ภ.พ.30 file (Manual mode — download for portal upload)
      tags: [Tax Filings]
      parameters:
        - name: year
          in: query
          required: true
          schema: { type: integer }
        - name: month
          in: query
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/xml: {}

  # ============================================================
  # ADMIN / SYSTEM
  # ============================================================
...
  /company-profile:
    get:
      summary: Get current tenant's company profile
      description: 'Any authenticated user can read (used by document headers).'
      tags: [Company Profile]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyProfile'
        '401':
          $ref: '#/components/responses/ErrorEnvelope401'
  /company-profile/soft:
    put:
      summary: Update soft fields (admin only)
      description: |
        Update editable contact/branding fields: trade_name, logo_url, phone,
        email, website, contact_name, bank_*. Requires `master.company.manage`.
      tags: [Company Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyProfileSoftUpdate'
      responses:
        '204':
          description: Updated
        '400':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelopeV1'
        '403':
          description: Missing master.company.manage scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelopeV1'
  /company-profile/company-info:
    put:
      summary: Edit full company info (super-admin) — founding identity + tax config + registered address
      description: |
        Super-admin (`master.company.manage`). Implements the previously-deferred /hard path: corrects
        the founding legal identity (legal name, 13-digit tax id, registration number, branch code,
        legal entity type), the VAT/tax config (§4.6) and the registered address in ONE call. Writes
        BOTH master.companies and company_profiles; every change is audited (tax_config_change +
        CompanyInfoChanged). §4.2-safe — posted tax invoices snapshot the supplier identity at
        post-time, so this affects FUTURE documents only.
      tags: [Company Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [legalName, taxId, branchCode, legalEntityType, province, postalCode]
              properties:
                legalName: { type: string }
                nameEn: { type: string, nullable: true }
                taxId: { type: string, description: 13-digit with valid checksum }
                registrationNumber: { type: string, nullable: true }
                legalEntityType: { type: string }
                branchCode: { type: string, description: 5 digits (00000 = head office) }
                vatRegistered: { type: boolean }
                vatRate: { type: number }
                pnd30SubmissionMode: { type: string, enum: [manual, auto] }
                vatRegisterDate: { type: string, format: date, nullable: true }
                building: { type: string, nullable: true }
                roomNo: { type: string, nullable: true }
                floor: { type: string, nullable: true }
                village: { type: string, nullable: true }
                houseNo: { type: string, nullable: true }
                moo: { type: string, nullable: true }
                soi: { type: string, nullable: true }
                street: { type: string, nullable: true }
                subdistrict: { type: string, nullable: true }
                district: { type: string, nullable: true }
                province: { type: string }
                postalCode: { type: string }
      responses:
        '204':
          description: Updated
        '400':
          description: Validation failure (tax id checksum, branch 5 digits, vat rate 0–1)
        '403':
          description: Missing master.company.manage scope
  /company-profile/hard:
    put:
      summary: Update hard fields (Phase 1 = 501 Not Implemented)
      description: |
        Hard fields (legal_name, tax_id, registration_number, registered_address_*,
        vat_registration_date, branch_code) are read-only via UI in Phase 1.
        Always returns 501 with explanation pointing to ภ.พ.09 workaround
        (file with revenue department + ops update DB manually with audit log).

        Phase 2 will add 2-person approval workflow + effective-date history
        + ภ.พ.09 attachment upload.
      tags: [Company Profile]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyProfileHardUpdate'
      responses:
        '501':
          description: 'Not implemented in Phase 1 — type=company_profile.hard_locked'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelopeV1'
        '403':
          description: Missing master.company.manage scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelopeV1'

  # ============================================================
  # WHT TYPES — Reactivate (Sprint 13f-P2; mirrors BU/Product Restore)
  # ============================================================
...
  /payroll/runs/{id}/sso/file:
    get:
      summary: สปส.1-10 e-Service text file (TIS-620 fixed-width 135 chars; เดือนตามงวดค่าจ้าง)
      tags: [Payroll]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        '200':
          description: Text file
          content:
            text/plain: {}
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Format: `key_<40-char-random>`. Issued via `POST /api-keys` (admin BFF).
        Plaintext returned once on create/rotate; only bcrypt hash + 16-char prefix stored.
        Scoped per-key (ScopesJson array of permission strings) + optional default_business_unit_id (auto-fill + lock per binding behavior).
        Required on all `/api/v1/*` endpoints (auth isolation: JWT not accepted on v1; ApiKey not accepted on root).
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'BFF cookie-based JWT for human users via /auth/login. Required on all root endpoints (auth isolation: ApiKey not accepted on root).'
    OAuth2:
      type: oauth2
      description: 'Reserved for Phase 2 client_credentials flow when first customer demands OAuth. API key sufficient for Phase 1.'
      flows:
        clientCredentials:
          tokenUrl: https://api.your-teas.example/oauth/token
          scopes:
            read: Read access
            write: Write access
            admin: Admin access
...
  responses:
...
    ErrorEnvelope401:
      description: 'Authentication failed. Body uses v1 envelope.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelopeV1'
  schemas:
...
    ErrorEnvelopeV1:
      type: object
      description: |
        Unified RFC-7807-style error envelope for ALL TEAS paths (root BFF
        + /api/v1/* + /api/proxy/*). Sprint 13d-P5 collapsed the prior
        root-vs-v1 split into this single shape via
        ValidationErrorEnvelopeMiddleware (transparent reshape).

        Stable error type catalog (urn:teas:error:*):
        - auth.unauthenticated | auth.forbidden
        - auth.missing_api_key | auth.invalid_api_key | auth.expired_api_key | auth.revoked_api_key | auth.scope_required
        - idempotency.required | idempotency.body_mismatch | idempotency.conflict
        - validation                                           (validation failures use fieldErrors[])
        - tenant.cross_tenant_access | tenant.required
        - business_unit.required | business_unit.locked_mismatch | business_unit.cross_bu_not_allowed_for_this_key
        - bu.duplicate | product.duplicate | wht_type.duplicate (business-rule uniqueness violations)
        - company_profile.hard_locked                          (PUT /company-profile/hard always returns 501)
        - period.closed | document.immutable_after_post
        - notfound                                             (resource id not in tenant scope)
        - internal                                             (uncaught 5xx)
        - rate_limit.exceeded (Phase 2)
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          example: 'urn:teas:error:validation'
          description: 'URN identifying the error class. Stable for programmatic handling.'
        title:
          type: string
          example: 'validation'
          description: 'i18n key (NOT human English). FE resolves to current locale.'
        detail:
          type: string
          example: 'Request validation failed (2 field(s)).'
          description: 'Optional human-readable detail in service origin locale. Mainly for logs/debugging.'
        status:
          type: integer
          example: 400
          description: 'Matches the HTTP status code.'
        fieldErrors:
          type: array
          description: 'Present for validation failures (400). Empty/absent for business-rule errors.'
          items:
            type: object
            required: [field, messages]
            properties:
              field:
                type: string
                example: 'code'
                description: 'camelCase field name matching request JSON shape (NOT C# Pascal DTO).'
              messages:
                type: array
                description: 'Array of i18n keys (typically validation.<rule>). FE resolves + can show inline beside the field.'
                items:
                  type: string
                  example: 'validation.required'
        trace_id: { type: string, example: 'abc123def456', description: 'From Activity.Current?.Id' }
        request_id: { type: string, example: 'req_xyz789', description: 'From HttpContext.TraceIdentifier' }
...
    CompanyProfile:
      type: object
      description: 'Tenant company profile. Hard fields are read-only via UI in Phase 1.'
      properties:
        companyId: { type: integer }
        # Hard fields — read-only via UI Phase 1
        legalName: { type: string, example: 'บริษัท แมนนวล เดโม จำกัด' }
        taxId: { type: string, example: '0000000000002' }
        registrationNumber: { type: string, example: '0000000000002' }
        registeredAddressLine1: { type: string }
        registeredAddressLine2: { type: string, nullable: true }
        registeredSubdistrict: { type: string }
        registeredDistrict: { type: string }
        registeredProvince: { type: string }
        registeredPostalCode: { type: string, example: '10110' }
        vatRegistrationDate: { type: string, format: date, nullable: true }
        branchCode: { type: string, example: '00000' }
        # Soft fields — admin-editable Phase 1
        tradeName: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        website: { type: string, nullable: true }
        contactName: { type: string, nullable: true }
        bankName: { type: string, nullable: true }
        bankAccountNo: { type: string, nullable: true }
        bankAccountName: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time }
        updatedByUserId: { type: integer, nullable: true }
    DefaultDocNotes:
      type: object
      description: |
        Per-document-kind default หมายเหตุ (doc-signature-and-foot-layout spec §G1). A
        CREATE-TIME FORM PREFILL ONLY — no document ever links back to this setting; changing
        it never edits any existing document's already-stored notes. A null/absent key means
        "no default" (the create form's Notes field opens blank, today's behaviour). Max 1000
        characters per value; a blank/whitespace value normalises to null.
      properties:
        quotation: { type: string, nullable: true, maxLength: 1000 }
        salesOrder: { type: string, nullable: true, maxLength: 1000 }
        deliveryOrder: { type: string, nullable: true, maxLength: 1000 }
        taxInvoice: { type: string, nullable: true, maxLength: 1000 }
        receipt: { type: string, nullable: true, maxLength: 1000 }
        billingNote: { type: string, nullable: true, maxLength: 1000 }
        creditNote: { type: string, nullable: true, maxLength: 1000 }
        debitNote: { type: string, nullable: true, maxLength: 1000 }
        purchaseOrder: { type: string, nullable: true, maxLength: 1000 }
        paymentVoucher: { type: string, nullable: true, maxLength: 1000 }
    CompanyProfileSoftUpdate:
      type: object
      description: |
        Editable fields only. All optional — omitted fields are unchanged.

        WARNING (defaultDocNotes only): this field is a WHOLE-JSONB-OVERWRITE, not a partial
        patch — omitting it (or sending fewer than all ten keys) CLEARS/blanks the stored
        per-doctype default notes. If the caller is only changing another soft field (e.g.
        tradeName) and wants to KEEP the existing defaultDocNotes, it must resend the complete
        current defaultDocNotes object in the same request (fetch it from GET /company-profile
        first). This mirrors the repo's existing whole-overwrite convention for this endpoint.
      properties:
        tradeName: { type: string, nullable: true }
        logoUrl: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        website: { type: string, nullable: true }
        contactName: { type: string, nullable: true }
        bankName: { type: string, nullable: true }
        bankAccountNo: { type: string, nullable: true }
        bankAccountName: { type: string, nullable: true }
        defaultDocNotes:
          allOf: [{ $ref: '#/components/schemas/DefaultDocNotes' }]
          nullable: true
    CompanyProfileHardUpdate:
      type: object
      description: |
        Hard-field update request body. Phase 1 endpoint always returns 501;
        this schema is documented for Phase 2 contract (when 2-person approval
        + ภ.พ.09 attachment is added). DO NOT call /company-profile/hard
        expecting success in Phase 1.
      properties:
        legalName: { type: string }
        taxId: { type: string }
        registrationNumber: { type: string }
        registeredAddressLine1: { type: string }
        registeredAddressLine2: { type: string, nullable: true }
        registeredSubdistrict: { type: string }
        registeredDistrict: { type: string }
        registeredProvince: { type: string }
        registeredPostalCode: { type: string }
        vatRegistrationDate: { type: string, format: date, nullable: true }
        branchCode: { type: string }
...